+++ dcrdminer/dcrdminer/Work.swift
@@ -9,13 +9,13 @@
 import Foundation
 
 class Work {
-    init(blockHeader: BlockHeader, targetDifficulty: Int64) {
+    init(blockHeader: BlockHeader, targetDifficulty: Int) {
         self.blockHeader = blockHeader
         self.targetDifficulty = targetDifficulty
         self.timestamp = Int64(NSDate().timeIntervalSince1970 * 1000)
     }
     
     var blockHeader: BlockHeader
-    let targetDifficulty: Int64
+    let targetDifficulty: Int
     let timestamp: Int64
 }
